🚚  formatting함수들 각 컴포넌트로 이동
